Move `*SpendingCondition` structs
