+++ stacks-codec/src/transaction.rs
@@ -25,12 +25,20 @@ use std::fmt::{self, Display};
 use std::io::{Read, Write};
 
 use serde::{Deserialize, Serialize};
-use stacks_common::address::AddressHashMode;
-use stacks_common::codec::{read_next, write_next, Error as codec_error, StacksMessageCodec};
-use stacks_common::types::chainstate::{ConsensusHash, StacksBlockId, StacksPublicKey};
-use stacks_common::types::StacksPublicKeyBuffer;
+use stacks_common::address::{
+    AddressHashMode, C32_ADDRESS_VERSION_MAINNET_MULTISIG, C32_ADDRESS_VERSION_MAINNET_SINGLESIG,
+    C32_ADDRESS_VERSION_TESTNET_MULTISIG, C32_ADDRESS_VERSION_TESTNET_SINGLESIG,
+};
+use stacks_common::codec::{
+    read_next, write_next, Error as codec_error, StacksMessageCodec, MAX_MESSAGE_LEN,
+};
+use stacks_common::types::chainstate::{
+    ConsensusHash, StacksAddress, StacksBlockId, StacksPrivateKey, StacksPublicKey, Txid,
+};
+use stacks_common::types::{PrivateKey, StacksEpochId, StacksPublicKeyBuffer};
 use stacks_common::util::hash::Hash160;
-use stacks_common::util::secp256k1::MessageSignature;
+use stacks_common::util::retry::BoundReader;
+use stacks_common::util::secp256k1::{MessageSignature, MESSAGE_SIGNATURE_ENCODED_SIZE};
 use stacks_common::{
     define_u8_enum, impl_array_hexstring_fmt, impl_array_newtype, impl_byte_array_message_codec,
     impl_byte_array_newtype, impl_byte_array_serde, impl_index_newtype,
@@ -694,5 +702,1048 @@ impl StacksMessageCodec for TransactionAuthField {
             }
         };
         Ok(field)
+    }
+}
+
+/// A structure that encodes enough state to authenticate
+/// a transaction's execution against a Stacks address.
+/// public_keys + signatures_required determines the Principal.
+/// nonce is the "check number" for the Principal.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct MultisigSpendingCondition {
+    pub hash_mode: MultisigHashMode,
+    pub signer: Hash160,
+    pub nonce: u64,  // nth authorization from this account
+    pub tx_fee: u64, // microSTX/compute rate offered by this account
+    pub fields: Vec<TransactionAuthField>,
+    pub signatures_required: u16,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct SinglesigSpendingCondition {
+    pub hash_mode: SinglesigHashMode,
+    pub signer: Hash160,
+    pub nonce: u64,  // nth authorization from this account
+    pub tx_fee: u64, // microSTX/compute rate offerred by this account
+    pub key_encoding: TransactionPublicKeyEncoding,
+    pub signature: MessageSignature,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct OrderIndependentMultisigSpendingCondition {
+    pub hash_mode: OrderIndependentMultisigHashMode,
+    pub signer: Hash160,
+    pub nonce: u64,  // nth authorization from this account
+    pub tx_fee: u64, // microSTX/compute rate offered by this account
+    pub fields: Vec<TransactionAuthField>,
+    pub signatures_required: u16,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub enum TransactionSpendingCondition {
+    Singlesig(SinglesigSpendingCondition),
+    Multisig(MultisigSpendingCondition),
+    OrderIndependentMultisig(OrderIndependentMultisigSpendingCondition),
+}
+
+impl StacksMessageCodec for MultisigSpendingCondition {
+    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
+        write_next(fd, &(self.hash_mode.clone() as u8))?;
+        write_next(fd, &self.signer)?;
+        write_next(fd, &self.nonce)?;
+        write_next(fd, &self.tx_fee)?;
+        write_next(fd, &self.fields)?;
+        write_next(fd, &self.signatures_required)?;
+        Ok(())
+    }
+
+    fn consensus_deserialize<R: Read>(
+        fd: &mut R,
+    ) -> Result<MultisigSpendingCondition, codec_error> {
+        let hash_mode_u8: u8 = read_next(fd)?;
+        let hash_mode = MultisigHashMode::from_u8(hash_mode_u8).ok_or(
+            codec_error::DeserializeError(format!(
+                "Failed to parse multisig spending condition: unknown hash mode {}",
+                hash_mode_u8
+            )),
+        )?;
+
+        let signer: Hash160 = read_next(fd)?;
+        let nonce: u64 = read_next(fd)?;
+        let tx_fee: u64 = read_next(fd)?;
+        let fields: Vec<TransactionAuthField> = {
+            let mut bound_read = BoundReader::from_reader(fd, MAX_MESSAGE_LEN as u64);
+            read_next(&mut bound_read)
+        }?;
+
+        let signatures_required: u16 = read_next(fd)?;
+
+        // read and decode _exactly_ num_signatures signature buffers
+        let mut num_sigs_given: u16 = 0;
+        let mut have_uncompressed = false;
+        for f in fields.iter() {
+            match *f {
+                TransactionAuthField::Signature(ref key_encoding, _) => {
+                    num_sigs_given =
+                        num_sigs_given
+                            .checked_add(1)
+                            .ok_or(codec_error::DeserializeError(
+                                "Failed to parse multisig spending condition: too many signatures"
+                                    .to_string(),
+                            ))?;
+                    if *key_encoding == TransactionPublicKeyEncoding::Uncompressed {
+                        have_uncompressed = true;
+                    }
+                }
+                TransactionAuthField::PublicKey(ref pubk) => {
+                    if !pubk.compressed() {
+                        have_uncompressed = true;
+                    }
+                }
+            };
+        }
+
+        // must be given the right number of signatures
+        if num_sigs_given != signatures_required {
+            return Err(codec_error::DeserializeError(format!(
+                "Failed to parse multisig spending condition: got {} sigs, expected {}",
+                num_sigs_given, signatures_required
+            )));
+        }
+
+        // must all be compressed if we're using P2WSH
+        if have_uncompressed && hash_mode == MultisigHashMode::P2WSH {
+            return Err(codec_error::DeserializeError(
+                "Failed to parse multisig spending condition: expected compressed keys only"
+                    .to_string(),
+            ));
+        }
+
+        Ok(MultisigSpendingCondition {
+            signer,
+            nonce,
+            tx_fee,
+            hash_mode,
+            fields,
+            signatures_required,
+        })
+    }
+}
+
+impl MultisigSpendingCondition {
+    pub fn push_signature(
+        &mut self,
+        key_encoding: TransactionPublicKeyEncoding,
+        signature: MessageSignature,
+    ) {
+        self.fields
+            .push(TransactionAuthField::Signature(key_encoding, signature));
+    }
+
+    pub fn push_public_key(&mut self, public_key: StacksPublicKey) {
+        self.fields
+            .push(TransactionAuthField::PublicKey(public_key));
+    }
+
+    pub fn pop_auth_field(&mut self) -> Option<TransactionAuthField> {
+        self.fields.pop()
+    }
+
+    pub fn address_mainnet(&self) -> StacksAddress {
+        StacksAddress::new(C32_ADDRESS_VERSION_MAINNET_MULTISIG, self.signer.clone())
+            .expect("FATAL: infallible: constant is not a valid address byte")
+    }
+
+    pub fn address_testnet(&self) -> StacksAddress {
+        StacksAddress::new(C32_ADDRESS_VERSION_TESTNET_MULTISIG, self.signer.clone())
+            .expect("FATAL: infallible: constant is not a valid address byte")
+    }
+
+    /// Authenticate a spending condition against an initial sighash.
+    /// In doing so, recover all public keys and verify that they hash to the signer
+    /// via the given hash mode.
+    pub fn verify(
+        &self,
+        initial_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+    ) -> Result<Txid, AuthError> {
+        let mut pubkeys = vec![];
+        let mut cur_sighash = initial_sighash.clone();
+        let mut num_sigs: u16 = 0;
+        let mut have_uncompressed = false;
+        for field in self.fields.iter() {
+            let pubkey = match field {
+                TransactionAuthField::PublicKey(ref pubkey) => {
+                    if !pubkey.compressed() {
+                        have_uncompressed = true;
+                    }
+                    pubkey.clone()
+                }
+                TransactionAuthField::Signature(ref pubkey_encoding, ref sigbuf) => {
+                    if *pubkey_encoding == TransactionPublicKeyEncoding::Uncompressed {
+                        have_uncompressed = true;
+                    }
+
+                    let (pubkey, next_sighash) = TransactionSpendingCondition::next_verification(
+                        &cur_sighash,
+                        cond_code,
+                        self.tx_fee,
+                        self.nonce,
+                        pubkey_encoding,
+                        sigbuf,
+                    )?;
+                    cur_sighash = next_sighash;
+                    num_sigs = num_sigs
+                        .checked_add(1)
+                        .ok_or(AuthError::VerifyingError("Too many signatures".to_string()))?;
+                    pubkey
+                }
+            };
+            pubkeys.push(pubkey);
+        }
+
+        if num_sigs != self.signatures_required {
+            return Err(AuthError::VerifyingError(
+                "Incorrect number of signatures".to_string(),
+            ));
+        }
+
+        if have_uncompressed && self.hash_mode == MultisigHashMode::P2WSH {
+            return Err(AuthError::VerifyingError(
+                "Uncompressed keys are not allowed in this hash mode".to_string(),
+            ));
+        }
+
+        let addr = StacksAddress::from_public_keys(
+            0,
+            &self.hash_mode.to_address_hash_mode(),
+            self.signatures_required as usize,
+            &pubkeys,
+        )
+        .ok_or_else(|| {
+            AuthError::VerifyingError("Failed to generate address from public keys".to_string())
+        })?;
+
+        if addr.bytes() != &self.signer {
+            return Err(AuthError::VerifyingError(format!(
+                "Signer hash does not equal hash of public key(s): {} != {}",
+                addr.bytes(),
+                self.signer
+            )));
+        }
+
+        Ok(cur_sighash)
+    }
+}
+
+impl StacksMessageCodec for OrderIndependentMultisigSpendingCondition {
+    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
+        write_next(fd, &(self.hash_mode.clone() as u8))?;
+        write_next(fd, &self.signer)?;
+        write_next(fd, &self.nonce)?;
+        write_next(fd, &self.tx_fee)?;
+        write_next(fd, &self.fields)?;
+        write_next(fd, &self.signatures_required)?;
+        Ok(())
+    }
+
+    fn consensus_deserialize<R: Read>(
+        fd: &mut R,
+    ) -> Result<OrderIndependentMultisigSpendingCondition, codec_error> {
+        let hash_mode_u8: u8 = read_next(fd)?;
+        let hash_mode = OrderIndependentMultisigHashMode::from_u8(hash_mode_u8).ok_or(
+            codec_error::DeserializeError(format!(
+                "Failed to parse multisig spending condition: unknown hash mode {}",
+                hash_mode_u8
+            )),
+        )?;
+
+        let signer: Hash160 = read_next(fd)?;
+        let nonce: u64 = read_next(fd)?;
+        let tx_fee: u64 = read_next(fd)?;
+        let fields: Vec<TransactionAuthField> = {
+            let mut bound_read = BoundReader::from_reader(fd, MAX_MESSAGE_LEN as u64);
+            read_next(&mut bound_read)
+        }?;
+
+        let signatures_required: u16 = read_next(fd)?;
+
+        // read and decode _exactly_ num_signatures signature buffers
+        let mut num_sigs_given: u16 = 0;
+        let mut have_uncompressed = false;
+        for f in fields.iter() {
+            match *f {
+                TransactionAuthField::Signature(ref key_encoding, _) => {
+                    num_sigs_given =
+                        num_sigs_given
+                            .checked_add(1)
+                            .ok_or(codec_error::DeserializeError(
+                                "Failed to parse order independent multisig spending condition: too many signatures"
+                                    .to_string(),
+                            ))?;
+                    if *key_encoding == TransactionPublicKeyEncoding::Uncompressed {
+                        have_uncompressed = true;
+                    }
+                }
+                TransactionAuthField::PublicKey(ref pubk) => {
+                    if !pubk.compressed() {
+                        have_uncompressed = true;
+                    }
+                }
+            };
+        }
+
+        // must be given the right number of signatures
+        if num_sigs_given < signatures_required {
+            let msg = format!(
+                "Failed to deserialize order independent multisig spending condition: got {num_sigs_given} sigs, expected at least {signatures_required}"
+            );
+            return Err(codec_error::DeserializeError(msg));
+        }
+
+        // must all be compressed if we're using P2WSH
+        if have_uncompressed && hash_mode == OrderIndependentMultisigHashMode::P2WSH {
+            let msg = "Failed to deserialize order independent multisig spending condition: expected compressed keys only".to_string();
+            return Err(codec_error::DeserializeError(msg));
+        }
+
+        Ok(OrderIndependentMultisigSpendingCondition {
+            signer,
+            nonce,
+            tx_fee,
+            hash_mode,
+            fields,
+            signatures_required,
+        })
+    }
+}
+
+impl OrderIndependentMultisigSpendingCondition {
+    pub fn push_signature(
+        &mut self,
+        key_encoding: TransactionPublicKeyEncoding,
+        signature: MessageSignature,
+    ) {
+        self.fields
+            .push(TransactionAuthField::Signature(key_encoding, signature));
+    }
+
+    pub fn push_public_key(&mut self, public_key: StacksPublicKey) {
+        self.fields
+            .push(TransactionAuthField::PublicKey(public_key));
+    }
+
+    pub fn pop_auth_field(&mut self) -> Option<TransactionAuthField> {
+        self.fields.pop()
+    }
+
+    pub fn address_mainnet(&self) -> StacksAddress {
+        StacksAddress::new(C32_ADDRESS_VERSION_MAINNET_MULTISIG, self.signer.clone())
+            .expect("FATAL: infallible: constant address byte is not supported")
+    }
+
+    pub fn address_testnet(&self) -> StacksAddress {
+        StacksAddress::new(C32_ADDRESS_VERSION_TESTNET_MULTISIG, self.signer.clone())
+            .expect("FATAL: infallible: constant address byte is not supported")
+    }
+
+    /// Authenticate a spending condition against an initial sighash.
+    /// In doing so, recover all public keys and verify that they hash to the signer
+    /// via the given hash mode.
+    pub fn verify(
+        &self,
+        initial_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+    ) -> Result<Txid, AuthError> {
+        let mut pubkeys = vec![];
+        let mut num_sigs: u16 = 0;
+        let mut have_uncompressed = false;
+        for field in self.fields.iter() {
+            let pubkey = match field {
+                TransactionAuthField::PublicKey(ref pubkey) => {
+                    if !pubkey.compressed() {
+                        have_uncompressed = true;
+                    }
+                    pubkey.clone()
+                }
+                TransactionAuthField::Signature(ref pubkey_encoding, ref sigbuf) => {
+                    if *pubkey_encoding == TransactionPublicKeyEncoding::Uncompressed {
+                        have_uncompressed = true;
+                    }
+
+                    let (pubkey, _next_sighash) = TransactionSpendingCondition::next_verification(
+                        initial_sighash,
+                        cond_code,
+                        self.tx_fee,
+                        self.nonce,
+                        pubkey_encoding,
+                        sigbuf,
+                    )?;
+                    num_sigs = num_sigs
+                        .checked_add(1)
+                        .ok_or(AuthError::VerifyingError("Too many signatures".to_string()))?;
+                    pubkey
+                }
+            };
+            pubkeys.push(pubkey);
+        }
+
+        if num_sigs < self.signatures_required {
+            return Err(AuthError::VerifyingError(format!(
+                "Not enough signatures. Got {num_sigs}, expected at least {req}",
+                req = self.signatures_required
+            )));
+        }
+
+        if have_uncompressed && self.hash_mode == OrderIndependentMultisigHashMode::P2WSH {
+            return Err(AuthError::VerifyingError(
+                "Uncompressed keys are not allowed in this hash mode".to_string(),
+            ));
+        }
+
+        let addr = StacksAddress::from_public_keys(
+            0,
+            &self.hash_mode.to_address_hash_mode(),
+            self.signatures_required as usize,
+            &pubkeys,
+        )
+        .ok_or_else(|| {
+            AuthError::VerifyingError("Failed to generate address from public keys".to_string())
+        })?;
+
+        if addr.bytes() != &self.signer {
+            return Err(AuthError::VerifyingError(format!(
+                "Signer hash does not equal hash of public key(s): {} != {}",
+                addr.bytes(),
+                self.signer
+            )));
+        }
+
+        Ok(initial_sighash.clone())
+    }
+}
+
+impl StacksMessageCodec for SinglesigSpendingCondition {
+    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
+        write_next(fd, &(self.hash_mode.clone() as u8))?;
+        write_next(fd, &self.signer)?;
+        write_next(fd, &self.nonce)?;
+        write_next(fd, &self.tx_fee)?;
+        write_next(fd, &(self.key_encoding as u8))?;
+        write_next(fd, &self.signature)?;
+        Ok(())
+    }
+
+    fn consensus_deserialize<R: Read>(
+        fd: &mut R,
+    ) -> Result<SinglesigSpendingCondition, codec_error> {
+        let hash_mode_u8: u8 = read_next(fd)?;
+        let hash_mode = SinglesigHashMode::from_u8(hash_mode_u8).ok_or(
+            codec_error::DeserializeError(format!(
+                "Failed to parse singlesig spending condition: unknown hash mode {}",
+                hash_mode_u8
+            )),
+        )?;
+
+        let signer: Hash160 = read_next(fd)?;
+        let nonce: u64 = read_next(fd)?;
+        let tx_fee: u64 = read_next(fd)?;
+
+        let key_encoding_u8: u8 = read_next(fd)?;
+        let key_encoding = TransactionPublicKeyEncoding::from_u8(key_encoding_u8).ok_or(
+            codec_error::DeserializeError(format!(
+                "Failed to parse singlesig spending condition: unknown key encoding {}",
+                key_encoding_u8
+            )),
+        )?;
+
+        let signature: MessageSignature = read_next(fd)?;
+
+        // sanity check -- must be compressed if we're using p2wpkh
+        if hash_mode == SinglesigHashMode::P2WPKH
+            && key_encoding != TransactionPublicKeyEncoding::Compressed
+        {
+            return Err(codec_error::DeserializeError("Failed to parse singlesig spending condition: incomaptible hash mode and key encoding".to_string()));
+        }
+
+        Ok(SinglesigSpendingCondition {
+            signer,
+            nonce,
+            tx_fee,
+            hash_mode,
+            key_encoding,
+            signature,
+        })
+    }
+}
+
+impl SinglesigSpendingCondition {
+    pub fn set_signature(&mut self, signature: MessageSignature) {
+        self.signature = signature;
+    }
+
+    pub fn pop_signature(&mut self) -> Option<TransactionAuthField> {
+        if self.signature == MessageSignature::empty() {
+            return None;
+        }
+
+        let ret = self.signature.clone();
+        self.signature = MessageSignature::empty();
+
+        return Some(TransactionAuthField::Signature(self.key_encoding, ret));
+    }
+
+    pub fn address_mainnet(&self) -> StacksAddress {
+        let version = match self.hash_mode {
+            SinglesigHashMode::P2PKH => C32_ADDRESS_VERSION_MAINNET_SINGLESIG,
+            SinglesigHashMode::P2WPKH => C32_ADDRESS_VERSION_MAINNET_MULTISIG,
+        };
+        StacksAddress::new(version, self.signer.clone())
+            .expect("FATAL: infallible: supported address constant is not valid")
+    }
+
+    pub fn address_testnet(&self) -> StacksAddress {
+        let version = match self.hash_mode {
+            SinglesigHashMode::P2PKH => C32_ADDRESS_VERSION_TESTNET_SINGLESIG,
+            SinglesigHashMode::P2WPKH => C32_ADDRESS_VERSION_TESTNET_MULTISIG,
+        };
+        StacksAddress::new(version, self.signer.clone())
+            .expect("FATAL: infallible: supported address constant is not valid")
+    }
+
+    /// Authenticate a spending condition against an initial sighash.
+    /// In doing so, recover all public keys and verify that they hash to the signer
+    /// via the given hash mode.
+    /// Returns the final sighash
+    pub fn verify(
+        &self,
+        initial_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+    ) -> Result<Txid, AuthError> {
+        let (pubkey, next_sighash) = TransactionSpendingCondition::next_verification(
+            initial_sighash,
+            cond_code,
+            self.tx_fee,
+            self.nonce,
+            &self.key_encoding,
+            &self.signature,
+        )?;
+
+        let addr = StacksAddress::from_public_keys(
+            0,
+            &self.hash_mode.to_address_hash_mode(),
+            1,
+            &vec![pubkey],
+        )
+        .ok_or_else(|| {
+            AuthError::VerifyingError("Failed to generate address from public key".to_string())
+        })?;
+
+        if addr.bytes() != &self.signer {
+            return Err(AuthError::VerifyingError(format!(
+                "Signer hash does not equal hash of public key(s): {} != {}",
+                addr.bytes(),
+                &self.signer
+            )));
+        }
+
+        Ok(next_sighash)
+    }
+}
+
+impl StacksMessageCodec for TransactionSpendingCondition {
+    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => {
+                data.consensus_serialize(fd)?;
+            }
+            TransactionSpendingCondition::Multisig(ref data) => {
+                data.consensus_serialize(fd)?;
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => {
+                data.consensus_serialize(fd)?;
+            }
+        }
+        Ok(())
+    }
+
+    fn consensus_deserialize<R: Read>(
+        fd: &mut R,
+    ) -> Result<TransactionSpendingCondition, codec_error> {
+        // peek the hash mode byte
+        let hash_mode_u8: u8 = read_next(fd)?;
+        let peek_buf = [hash_mode_u8];
+        let mut rrd = peek_buf.chain(fd);
+        let cond = {
+            if SinglesigHashMode::from_u8(hash_mode_u8).is_some() {
+                let cond = SinglesigSpendingCondition::consensus_deserialize(&mut rrd)?;
+                TransactionSpendingCondition::Singlesig(cond)
+            } else if MultisigHashMode::from_u8(hash_mode_u8).is_some() {
+                let cond = MultisigSpendingCondition::consensus_deserialize(&mut rrd)?;
+                TransactionSpendingCondition::Multisig(cond)
+            } else if OrderIndependentMultisigHashMode::from_u8(hash_mode_u8).is_some() {
+                let cond =
+                    OrderIndependentMultisigSpendingCondition::consensus_deserialize(&mut rrd)?;
+                TransactionSpendingCondition::OrderIndependentMultisig(cond)
+            } else {
+                return Err(codec_error::DeserializeError(format!(
+                    "Failed to parse spending condition: invalid hash mode {}",
+                    hash_mode_u8
+                )));
+            }
+        };
+
+        Ok(cond)
+    }
+}
+
+impl TransactionSpendingCondition {
+    pub fn new_singlesig_p2pkh(pubkey: StacksPublicKey) -> Option<TransactionSpendingCondition> {
+        let key_encoding = if pubkey.compressed() {
+            TransactionPublicKeyEncoding::Compressed
+        } else {
+            TransactionPublicKeyEncoding::Uncompressed
+        };
+        let signer_addr =
+            StacksAddress::from_public_keys(0, &AddressHashMode::SerializeP2PKH, 1, &vec![pubkey])?;
+
+        Some(TransactionSpendingCondition::Singlesig(
+            SinglesigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: SinglesigHashMode::P2PKH,
+                key_encoding,
+                signature: MessageSignature::empty(),
+            },
+        ))
+    }
+
+    pub fn new_singlesig_p2wpkh(pubkey: StacksPublicKey) -> Option<TransactionSpendingCondition> {
+        let signer_addr = StacksAddress::from_public_keys(
+            0,
+            &AddressHashMode::SerializeP2WPKH,
+            1,
+            &vec![pubkey],
+        )?;
+
+        Some(TransactionSpendingCondition::Singlesig(
+            SinglesigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: SinglesigHashMode::P2WPKH,
+                key_encoding: TransactionPublicKeyEncoding::Compressed,
+                signature: MessageSignature::empty(),
+            },
+        ))
+    }
+
+    pub fn new_multisig_p2sh(
+        num_sigs: u16,
+        pubkeys: Vec<StacksPublicKey>,
+    ) -> Option<TransactionSpendingCondition> {
+        let signer_addr = StacksAddress::from_public_keys(
+            0,
+            &AddressHashMode::SerializeP2SH,
+            usize::from(num_sigs),
+            &pubkeys,
+        )?;
+
+        Some(TransactionSpendingCondition::Multisig(
+            MultisigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: MultisigHashMode::P2SH,
+                fields: vec![],
+                signatures_required: num_sigs,
+            },
+        ))
+    }
+
+    pub fn new_multisig_order_independent_p2sh(
+        num_sigs: u16,
+        pubkeys: Vec<StacksPublicKey>,
+    ) -> Option<TransactionSpendingCondition> {
+        let signer_addr = StacksAddress::from_public_keys(
+            0,
+            &AddressHashMode::SerializeP2SH,
+            usize::from(num_sigs),
+            &pubkeys,
+        )?;
+
+        Some(TransactionSpendingCondition::OrderIndependentMultisig(
+            OrderIndependentMultisigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: OrderIndependentMultisigHashMode::P2SH,
+                fields: vec![],
+                signatures_required: num_sigs,
+            },
+        ))
+    }
+
+    pub fn new_multisig_order_independent_p2wsh(
+        num_sigs: u16,
+        pubkeys: Vec<StacksPublicKey>,
+    ) -> Option<TransactionSpendingCondition> {
+        let signer_addr = StacksAddress::from_public_keys(
+            0,
+            &AddressHashMode::SerializeP2WSH,
+            usize::from(num_sigs),
+            &pubkeys,
+        )?;
+
+        Some(TransactionSpendingCondition::OrderIndependentMultisig(
+            OrderIndependentMultisigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: OrderIndependentMultisigHashMode::P2WSH,
+                fields: vec![],
+                signatures_required: num_sigs,
+            },
+        ))
+    }
+
+    pub fn new_multisig_p2wsh(
+        num_sigs: u16,
+        pubkeys: Vec<StacksPublicKey>,
+    ) -> Option<TransactionSpendingCondition> {
+        let signer_addr = StacksAddress::from_public_keys(
+            0,
+            &AddressHashMode::SerializeP2WSH,
+            usize::from(num_sigs),
+            &pubkeys,
+        )?;
+
+        Some(TransactionSpendingCondition::Multisig(
+            MultisigSpendingCondition {
+                signer: signer_addr.destruct().1,
+                nonce: 0,
+                tx_fee: 0,
+                hash_mode: MultisigHashMode::P2WSH,
+                fields: vec![],
+                signatures_required: num_sigs,
+            },
+        ))
+    }
+
+    /// When committing to the fact that a transaction is sponsored, the origin doesn't know
+    /// anything else.  Instead, it commits to this sentinel value as its sponsor.
+    /// It is intractable to calculate a private key that could generate this.
+    pub fn new_initial_sighash() -> TransactionSpendingCondition {
+        TransactionSpendingCondition::Singlesig(SinglesigSpendingCondition {
+            signer: Hash160([0u8; 20]),
+            nonce: 0,
+            tx_fee: 0,
+            hash_mode: SinglesigHashMode::P2PKH,
+            key_encoding: TransactionPublicKeyEncoding::Compressed,
+            signature: MessageSignature::empty(),
+        })
+    }
+
+    pub fn num_signatures(&self) -> u16 {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => {
+                if data.signature != MessageSignature::empty() {
+                    1
+                } else {
+                    0
+                }
+            }
+            TransactionSpendingCondition::Multisig(ref data) => {
+                let mut num_sigs: u16 = 0;
+                for field in data.fields.iter() {
+                    if field.is_signature() {
+                        num_sigs = num_sigs
+                            .checked_add(1)
+                            .expect("Unreasonable amount of signatures"); // something is seriously wrong if this fails
+                    }
+                }
+                num_sigs
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => {
+                let mut num_sigs: u16 = 0;
+                for field in data.fields.iter() {
+                    if field.is_signature() {
+                        num_sigs = num_sigs
+                            .checked_add(1)
+                            .expect("Unreasonable amount of signatures"); // something is seriously wrong if this fails
+                    }
+                }
+                num_sigs
+            }
+        }
+    }
+
+    pub fn signatures_required(&self) -> u16 {
+        match *self {
+            TransactionSpendingCondition::Singlesig(_) => 1,
+            TransactionSpendingCondition::Multisig(ref multisig_data) => {
+                multisig_data.signatures_required
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref multisig_data) => {
+                multisig_data.signatures_required
+            }
+        }
+    }
+
+    pub fn nonce(&self) -> u64 {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => data.nonce,
+            TransactionSpendingCondition::Multisig(ref data) => data.nonce,
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => data.nonce,
+        }
+    }
+
+    pub fn tx_fee(&self) -> u64 {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => data.tx_fee,
+            TransactionSpendingCondition::Multisig(ref data) => data.tx_fee,
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => data.tx_fee,
+        }
+    }
+
+    pub fn set_nonce(&mut self, n: u64) {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref mut singlesig_data) => {
+                singlesig_data.nonce = n;
+            }
+            TransactionSpendingCondition::Multisig(ref mut multisig_data) => {
+                multisig_data.nonce = n;
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut multisig_data) => {
+                multisig_data.nonce = n;
+            }
+        }
+    }
+
+    pub fn set_tx_fee(&mut self, tx_fee: u64) {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref mut singlesig_data) => {
+                singlesig_data.tx_fee = tx_fee;
+            }
+            TransactionSpendingCondition::Multisig(ref mut multisig_data) => {
+                multisig_data.tx_fee = tx_fee;
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut multisig_data) => {
+                multisig_data.tx_fee = tx_fee;
+            }
+        }
+    }
+
+    pub fn get_tx_fee(&self) -> u64 {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref singlesig_data) => singlesig_data.tx_fee,
+            TransactionSpendingCondition::Multisig(ref multisig_data) => multisig_data.tx_fee,
+            TransactionSpendingCondition::OrderIndependentMultisig(ref multisig_data) => {
+                multisig_data.tx_fee
+            }
+        }
+    }
+
+    /// Get the mainnet account address of the spending condition
+    pub fn address_mainnet(&self) -> StacksAddress {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => data.address_mainnet(),
+            TransactionSpendingCondition::Multisig(ref data) => data.address_mainnet(),
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => {
+                data.address_mainnet()
+            }
+        }
+    }
+
+    /// Get the mainnet account address of the spending condition
+    pub fn address_testnet(&self) -> StacksAddress {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => data.address_testnet(),
+            TransactionSpendingCondition::Multisig(ref data) => data.address_testnet(),
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => {
+                data.address_testnet()
+            }
+        }
+    }
+
+    /// Get the address for an account, given the network flag
+    pub fn get_address(&self, mainnet: bool) -> StacksAddress {
+        if mainnet {
+            self.address_mainnet()
+        } else {
+            self.address_testnet()
+        }
+    }
+
+    /// Clear fee rate, nonces, signatures, and public keys
+    pub fn clear(&mut self) {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref mut singlesig_data) => {
+                singlesig_data.tx_fee = 0;
+                singlesig_data.nonce = 0;
+                singlesig_data.signature = MessageSignature::empty();
+            }
+            TransactionSpendingCondition::Multisig(ref mut multisig_data) => {
+                multisig_data.tx_fee = 0;
+                multisig_data.nonce = 0;
+                multisig_data.fields.clear();
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut multisig_data) => {
+                multisig_data.tx_fee = 0;
+                multisig_data.nonce = 0;
+                multisig_data.fields.clear();
+            }
+        }
+    }
+
+    pub fn make_sighash_presign(
+        cur_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+        tx_fee: u64,
+        nonce: u64,
+    ) -> Txid {
+        // new hash combines the previous hash and all the new data this signature will add.  This
+        // includes:
+        // * the previous hash
+        // * the auth flag
+        // * the fee rate (big-endian 8-byte number)
+        // * nonce (big-endian 8-byte number)
+        let new_tx_hash_bits_len = 32 + 1 + 8 + 8;
+        let mut new_tx_hash_bits = Vec::with_capacity(new_tx_hash_bits_len as usize);
+
+        new_tx_hash_bits.extend_from_slice(cur_sighash.as_bytes());
+        new_tx_hash_bits.extend_from_slice(&[*cond_code as u8]);
+        new_tx_hash_bits.extend_from_slice(&tx_fee.to_be_bytes());
+        new_tx_hash_bits.extend_from_slice(&nonce.to_be_bytes());
+
+        assert!(new_tx_hash_bits.len() == new_tx_hash_bits_len as usize);
+
+        let next_sighash = Txid::from_sighash_bytes(&new_tx_hash_bits);
+        next_sighash
+    }
+
+    pub fn make_sighash_postsign(
+        cur_sighash: &Txid,
+        pubkey: &StacksPublicKey,
+        sig: &MessageSignature,
+    ) -> Txid {
+        // new hash combines the previous hash and all the new data this signature will add.  This
+        // includes:
+        // * the public key compression flag
+        // * the signature
+        let new_tx_hash_bits_len = 32 + 1 + MESSAGE_SIGNATURE_ENCODED_SIZE;
+        let mut new_tx_hash_bits = Vec::with_capacity(new_tx_hash_bits_len as usize);
+        let pubkey_encoding = if pubkey.compressed() {
+            TransactionPublicKeyEncoding::Compressed
+        } else {
+            TransactionPublicKeyEncoding::Uncompressed
+        };
+
+        new_tx_hash_bits.extend_from_slice(cur_sighash.as_bytes());
+        new_tx_hash_bits.extend_from_slice(&[pubkey_encoding as u8]);
+        new_tx_hash_bits.extend_from_slice(sig.as_bytes());
+
+        assert!(new_tx_hash_bits.len() == new_tx_hash_bits_len as usize);
+
+        let next_sighash = Txid::from_sighash_bytes(&new_tx_hash_bits);
+        next_sighash
+    }
+
+    /// Linear-complexity signing algorithm -- we sign a rolling hash over all data committed to by
+    /// the previous signer (instead of naively re-serializing the transaction each time), as well
+    /// as over new data provided by this key (excluding its own public key or signature, which
+    /// are authenticated by the spending condition's key hash).
+    /// Calculates and returns the next signature and sighash, which the subsequent private key
+    /// must sign.
+    pub fn next_signature(
+        cur_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+        tx_fee: u64,
+        nonce: u64,
+        privk: &StacksPrivateKey,
+    ) -> Result<(MessageSignature, Txid), AuthError> {
+        let sighash_presign = TransactionSpendingCondition::make_sighash_presign(
+            cur_sighash,
+            cond_code,
+            tx_fee,
+            nonce,
+        );
+
+        // sign the current hash
+        let sig = privk
+            .sign(sighash_presign.as_bytes())
+            .map_err(|se| AuthError::SigningError(se.to_string()))?;
+
+        let pubk = StacksPublicKey::from_private(privk);
+        let next_sighash =
+            TransactionSpendingCondition::make_sighash_postsign(&sighash_presign, &pubk, &sig);
+
+        Ok((sig, next_sighash))
+    }
+
+    /// Linear-complexity verifying algorithm -- we verify a rolling hash over all data committed
+    /// to by order of signers (instead of re-serializing the tranasction each time).
+    /// Calculates the next sighash and public key, which the next verifier must verify.
+    /// Used by StacksTransaction::verify*
+    pub fn next_verification(
+        cur_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+        tx_fee: u64,
+        nonce: u64,
+        key_encoding: &TransactionPublicKeyEncoding,
+        sig: &MessageSignature,
+    ) -> Result<(StacksPublicKey, Txid), AuthError> {
+        let sighash_presign = TransactionSpendingCondition::make_sighash_presign(
+            cur_sighash,
+            cond_code,
+            tx_fee,
+            nonce,
+        );
+
+        // verify the current signature
+        let mut pubk = StacksPublicKey::recover_to_pubkey(sighash_presign.as_bytes(), sig)
+            .map_err(|ve| AuthError::VerifyingError(ve.to_string()))?;
+
+        match key_encoding {
+            TransactionPublicKeyEncoding::Compressed => pubk.set_compressed(true),
+            TransactionPublicKeyEncoding::Uncompressed => pubk.set_compressed(false),
+        };
+
+        // what's the next sighash going to be?
+        let next_sighash =
+            TransactionSpendingCondition::make_sighash_postsign(&sighash_presign, &pubk, sig);
+        Ok((pubk, next_sighash))
+    }
+
+    /// Verify all signatures
+    pub fn verify(
+        &self,
+        initial_sighash: &Txid,
+        cond_code: &TransactionAuthFlags,
+    ) -> Result<Txid, AuthError> {
+        match *self {
+            TransactionSpendingCondition::Singlesig(ref data) => {
+                data.verify(initial_sighash, cond_code)
+            }
+            TransactionSpendingCondition::Multisig(ref data) => {
+                data.verify(initial_sighash, cond_code)
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref data) => {
+                data.verify(initial_sighash, cond_code)
+            }
+        }
+    }
+
+    /// Checks if this TransactionSpendingCondition is supported in the passed epoch
+    /// OrderIndependent multisig is not supported before epoch 3.0
+    pub fn is_supported_in_epoch(&self, epoch_id: StacksEpochId) -> bool {
+        match self {
+            TransactionSpendingCondition::Singlesig(..)
+            | TransactionSpendingCondition::Multisig(..) => true,
+            TransactionSpendingCondition::OrderIndependentMultisig(..) => {
+                epoch_id >= StacksEpochId::Epoch30
+            }
+        }
     }
 }
