+++ stackslib/src/chainstate/stacks/mod.rs
@@ -384,47 +384,6 @@ impl Error {
     }
 }
 
-/// A structure that encodes enough state to authenticate
-/// a transaction's execution against a Stacks address.
-/// public_keys + signatures_required determines the Principal.
-/// nonce is the "check number" for the Principal.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct MultisigSpendingCondition {
-    pub hash_mode: MultisigHashMode,
-    pub signer: Hash160,
-    pub nonce: u64,  // nth authorization from this account
-    pub tx_fee: u64, // microSTX/compute rate offered by this account
-    pub fields: Vec<TransactionAuthField>,
-    pub signatures_required: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct SinglesigSpendingCondition {
-    pub hash_mode: SinglesigHashMode,
-    pub signer: Hash160,
-    pub nonce: u64,  // nth authorization from this account
-    pub tx_fee: u64, // microSTX/compute rate offerred by this account
-    pub key_encoding: TransactionPublicKeyEncoding,
-    pub signature: MessageSignature,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct OrderIndependentMultisigSpendingCondition {
-    pub hash_mode: OrderIndependentMultisigHashMode,
-    pub signer: Hash160,
-    pub nonce: u64,  // nth authorization from this account
-    pub tx_fee: u64, // microSTX/compute rate offered by this account
-    pub fields: Vec<TransactionAuthField>,
-    pub signatures_required: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum TransactionSpendingCondition {
-    Singlesig(SinglesigSpendingCondition),
-    Multisig(MultisigSpendingCondition),
-    OrderIndependentMultisig(OrderIndependentMultisigSpendingCondition),
-}
-
 /// Types of transaction authorizations
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub enum TransactionAuth {
@@ -457,11 +416,12 @@ pub struct TransactionSmartContract {
 
 pub use stacks_codec::transaction::{
     AssetInfoID, AuthError, CoinbasePayload, FungibleConditionCode, MultisigHashMode,
-    OrderIndependentMultisigHashMode, PostConditionPrincipalID, SinglesigHashMode,
-    TenureChangeCause, TenureChangeError, TenureChangePayload, TokenTransferMemo,
-    TransactionAnchorMode, TransactionAuthField, TransactionAuthFieldID, TransactionAuthFlags,
-    TransactionPayloadID, TransactionPostConditionMode, TransactionPublicKeyEncoding,
-    TransactionVersion,
+    MultisigSpendingCondition, OrderIndependentMultisigHashMode,
+    OrderIndependentMultisigSpendingCondition, PostConditionPrincipalID, SinglesigHashMode,
+    SinglesigSpendingCondition, TenureChangeCause, TenureChangeError, TenureChangePayload,
+    TokenTransferMemo, TransactionAnchorMode, TransactionAuthField, TransactionAuthFieldID,
+    TransactionAuthFlags, TransactionPayloadID, TransactionPostConditionMode,
+    TransactionPublicKeyEncoding, TransactionSpendingCondition, TransactionVersion,
 };
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
